Revert "extras/keymap: Fix case matching for Micro-Star"
This reverts commit 66bf63c05cdc4e9b09818aa5fab0b9d319a1c91c.

Further debugging in https://launchpad.net/bugs/178860 showed that for some
weird reason the correct key codes already come out of the "Video Bus" input
device, and the previous commit would cause them to appear a second time
through the standard keyboard device.

This is a kernel bug in the end, but let's not break working things
prematurely.