rules: Add symlink rule for virtio ports
virtio ports spawned by the virtio_console.c driver can have 'names'
assigned to them by hosts. The ports are distinguishable using these
names. Make a rule to create a symlink to the chardev associated for a
port with a name.

The symlink created is:
    /dev/virtio-ports/org.libvirt.console0 -> /dev/vport0p0

if the first port for the first device was given a name of
'org.libvirt.console0'.

Signed-off-by: Amit Shah <amit.shah@redhat.com>