Fix wlan key on Inspiron 1010 & 1110
This fixes wlan key on Inspirion 1010 & 1110.

This patch depends on previous patches sent.

The issue with all of these is that they were all Dell mini & it wasn't
noticed till recent that they all did not follow the standard that the
rest of Dell machines follow.

Also to note while this fixes the wlan key sending the proper key press,
work is still needed at the kernel level for complete support.

This is the last patch all the Dell minis have been verified to all have
this issue, and are now covered.

Signed-off-by: Jerone Young <jerone.young@canonical.com>
Signed-off-by: Martin Pitt <martin.pitt@ubuntu.com>