man: extend documentation on systemctl isolate
Lennart has convinced me that it's more helpful to participate than to sit
on the sidelines and complain. So, hello everyone.

I'm starting by giving up the battle to change the systemctl "isolate"
command to "switch-to". Can't win them all. :) I've got a suggested patch
to expand the documentation a bit, hopefully making it more clear to new
systemd users.

Is there an easy way to list all units where AllowIsolate is enabled? That
should be included alongside this, I think.