util: Add welcome message for Ubuntu
The approved method for determining the installed release of
Ubuntu is to execute "lsb_release".  However, this is in /usr/bin and
is implemented in python so it is not safe to execute at this early
stage of booting.  This code parses /etc/lsb-release which is
where "lsb_release" looks for the information.