vconsole: support additional keymap for toggling layouts
At least Mandriva offers configuring characters to toggle
keyboard layout independently from main keymap. This functions
much the same as XkbOptions for X11 and actually is configured
together. The patch adds support for additional keymap,
KEYMAP_TOGGLE, to /etc/vconsole.conf, that is intended to be
used for the same purpose.