path: fix watching the root directory
If you have a path unit with:
  DirectoryNotEmpty=/foo
and "/foo" does not exist, creating it later escapes the attention of systemd.

When adding watches for parent directories, systemd never adds one for the root
directory. It attempts to add a watch for an empty string instead, which fails.

If the path is "/", we must not trim the slash.