add org.freedesktop.DBus.Properies.Set method
The patch adds framework for processing Set requests on properties
and as first consumers allows setting of LogTarget and LogLevel on
interface org.freedesktop.systemd1.Manager of org.freedesktop.systemd1.

The code should be generic enough to allow processing of any
property on other objects and interfaces as well.

No systemctl visible interface is added so far. The LogTarget and
LogLevel are meant for debugging in the first place; user interface
can be added if other usage emerge.

Set on systemwide systemd is restricted to root; I am not sure
how session level access is controlled.