keymap: Support Dell Latitude XT2 tablet-mode navigation keys
The XT2 has a rocker (up/down/enter) and back button on the side
in tablet mode, none of which work currently. Add entries for
these keys.

There is some overlap here with scan codes used in other Dell
models, so these buttons are put in a new file specific to this
model.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Martin Pitt <martin.pitt@ubuntu.com>