drop hwclock-save.service
We don't want to fiddle around changing the RTC, not on bootup, not
on shutdown.

If we don't run NTP, we have absolutely no clue what's the current
time to store in the RTC. If we run NTP, the kernel syncs the system
time every 11 minutes to the RTC.

Especially in multi-boot environents we must not call hwclock(8)
which tries to be smart with calculating/storing/applying drifts
and such.

Live-CDs must never touch the RTC, because we don't know if it is
running in UTC or locatime.