tmpfiles: don't exit with an error code if we cannot access all files
This is just to avoid confusion if people use stuff like FUSE file
systems or SELinux which might disallow access to files even if tmpfiles
is running as root.

We still log away if we cannot access a file, but we do not return a
failure exit code in the end.