man: list specifiers in a table
It was noted on IRC that the current layout is not necessarily the
easiest to parse by humans. I personnally struggled understanding it the
first time I read it.

Hopefully, this change makes it easier to understand.

Note: I only reformatted the information that was already present in the
documentation, I didn't add anything.