service: warn if a service fails to write its PID file
Warn if a service promises to write a PID file (using 'PIDFile=' in the
unit file or '# pidfile:' in SysV header), but fails to keep the
promise.

This warning will likely trigger also for the forking services with a
racy daemonization, which exit the original process before the PID file
is written.